+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-11T12:15:32.022Z",
+    "2026-09-11T14:43:29.884Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {