+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1764,7 +1764,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M286A5GHZPBP85SW9EKYMK5D"),
+  _id: z.string().default("tool_01M28ES5AJ4E7ADQRTRHJSVTFP"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -1959,7 +1959,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M286A5GGVC0BE0SH0GV42DJ3"),
+  _id: z.string().default("tool_01M28ES5AH10CR533V896FW0PY"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2054,7 +2054,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M286A5GGTHHTHVPYRGH67RYB"),
+  _id: z.string().default("tool_01M28ES5AG7BXTT0QD62ZHX66E"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2157,7 +2157,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01M286A5GF130RZN94AAANDNMS"),
+  _id: z.string().default("tool_01M28ES5AF8ZQQ37BSVYWHM5D9"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),